refactor(aarch64): extract hermit-entry-specific start code
