+++ src/arch/aarch64/mod.rs
@@ -1,2 +1,4 @@
 pub mod kernel;
 pub mod mm;
+#[cfg(target_os = "none")]
+pub mod start;
