+++ src/arch/aarch64/start/mod.rs
@@ -1,0 +1,3 @@
+pub mod hermit_entry;
+#[cfg(feature = "smp")]
+pub mod smp;
